Add Docker Compose configuration for schwifty-server stack
This commit adds the Docker Compose configuration for the schwifty-server stack. It includes services for cloudflared, ibragimov.by-currency, ruslan.ibragimov.by, heapy.io-kotlin_jobs, and heapy.io. Each service has specific resource limits for CPUs and memory. The services are declared in the schwifty network and have dependencies on each other. The containers use the journald logging driver and are set to restart always. The stop grace period and memory swappiness are also defined.